+++ src/Base/Element.php
@@ -131,13 +131,30 @@ abstract class Element
         foreach($classes as $class) {           
             $media = explode(':', $class)[0];
             if (in_array($media, array_keys($medias))) {
-                $result[$media][] = strings($class)->replace($media . ':', '')->toString();
+                $className = strings($class)->replace($media . ':', '')->toString();
+                $result[$media][$className] = $className;
                 continue;
             }
-            $result['global'][] = $class;
+            $result['global'][$class] = $class;
         }
 
         return collection($result);
+    }
+
+    /**
+     * Call the given callback.
+     *
+     * @param callable $callback Callback function.
+     *
+     * @return self Returns instance of the Element class.
+     *
+     * @access public
+     */
+    public function pipe(callable $callback): self
+    {
+        $callback($this);
+
+        return $this;
     }
 
     /**
@@ -221,6 +238,38 @@ abstract class Element
     {
         $this->classes = $this->makeClasses($classes);
 
+        return $this;
+    }
+
+    /**
+     * Add element classes.
+     *
+     * @param string $classes Element classes.
+     *
+     * @return self Returns instance of the Element class.
+     *
+     * @access public
+     */
+    public function addClasses(string $classes = ''): self
+    {
+        $this->classes->replace($this->makeClasses($classes), true);
+
+        return $this;
+    }
+
+    /**
+     * Remove element classes.
+     *
+     * @param string $classes Element classes.
+     *
+     * @return self Returns instance of the Element class.
+     *
+     * @access public
+     */
+    public function removeClasses(string $classes = ''): self
+    {
+        $this->classes->delete($this->makeClasses($classes)->dot()->divide()[0]);
+       
         return $this;
     }